requirements.txt
fastapi
uvicorn
python-multipart
streamlit
torch
torchvision
pillow
numpy
matplotlib
requests
opencv-python-headless